refactor: isolate table caption logic and table transformer
