+++ EMS/admin-ui-bundle/assets/src/core/components/tiptap/module/table.ts
@@ -1,13 +1,25 @@
-import { Node, mergeAttributes, Editor } from '@tiptap/core'
+import { Node, Editor } from '@tiptap/core'
 import { Table, TableCell, TableHeader, TableRow } from '@tiptap/extension-table'
 import IconTable from '@tabler/icons/outline/table.svg?raw'
-import { HtmlTransform, ContextMenuItem, TiptapModule } from '../types.ts'
+import { ContextMenuItem, TiptapModule } from '../types.ts'
 import { Dialog } from '../../dialog.ts'
 import { TiptapEditor } from '../editor.ts'
 
+import {
+    TableCaption,
+    tableCaptionHtmlTransform,
+    TableFigure,
+    updateCaption
+} from '../table/tableCaption.ts'
+
+import { tableCleanHtmlTransform } from '../table/table.ts'
+
 export const tableModule: TiptapModule = {
     extensions: getExtensions(),
-    htmlTransforms: getHtmlTransforms(),
+    htmlTransforms: [
+        tableCaptionHtmlTransform,
+        tableCleanHtmlTransform
+    ],
     toolbarGroup: 'insert',
     toolbar: [
         {
@@ -58,50 +70,6 @@ function getExtensions(): Node[] {
             }
         }
     })
-    const TableFigure = Node.create({
-        name: 'tableFigure',
-        group: 'block',
-        content: 'tableCaption? table',
-        isolating: true,
-
-        parseHTML() {
-            return [{ tag: 'figure[data-type="table"]' }]
-        },
-
-        renderHTML({ HTMLAttributes }) {
-            return ['figure', mergeAttributes(HTMLAttributes, { 'data-type': 'table' }), 0]
-        }
-    })
-    const TableCaption = Node.create({
-        name: 'tableCaption',
-        content: 'inline*',
-        defining: true,
-
-        parseHTML() {
-            return [{ tag: 'figcaption' }]
-        },
-
-        renderHTML({ HTMLAttributes }) {
-            return ['figcaption', mergeAttributes(HTMLAttributes), 0]
-        },
-
-        addKeyboardShortcuts() {
-            return {
-                Backspace: () => {
-                    const { $from, empty } = this.editor.state.selection
-                    if (!empty) return false
-                    if ($from.parent.type.name !== this.name) return false
-                    if ($from.parent.content.size > 0) return false
-
-                    return this.editor
-                        .chain()
-                        .deleteRange({ from: $from.before(), to: $from.after() })
-                        .focus()
-                        .run()
-                }
-            }
-        }
-    })
 
     return [
         CustomTable.configure({ resizable: false, allowTableNodeSelection: true }),
@@ -110,78 +78,6 @@ function getExtensions(): Node[] {
         TableHeader,
         TableFigure,
         TableCaption
-    ]
-}
-
-function getHtmlTransforms(): HtmlTransform[] {
-    return [
-        {
-            name: 'tableCaption',
-            toEditor(doc) {
-                doc.querySelectorAll('table').forEach((table) => {
-                    const caption = table.querySelector(':scope > caption')
-                    if (!caption) return
-                    const figure = doc.createElement('figure')
-                    figure.setAttribute('data-type', 'table')
-                    const figcaption = doc.createElement('figcaption')
-                    figcaption.innerHTML = caption.innerHTML
-                    caption.remove()
-                    table.replaceWith(figure)
-                    figure.appendChild(figcaption)
-                    figure.appendChild(table)
-                })
-            },
-            toOutput(doc) {
-                doc.querySelectorAll('figure[data-type="table"]').forEach((fig) => {
-                    const figcaption = fig.querySelector(':scope > figcaption')
-                    const table = fig.querySelector(':scope > table')
-                    if (!table) {
-                        fig.remove()
-                        return
-                    }
-                    if (figcaption) {
-                        const caption = doc.createElement('caption')
-                        caption.innerHTML = figcaption.innerHTML
-                        table.insertBefore(caption, table.firstChild)
-                    }
-                    fig.replaceWith(table)
-                })
-            }
-        },
-        {
-            name: 'tableCleanup',
-            toEditor(doc) {
-                doc.querySelectorAll('table').forEach((table) => {
-                    const style = table.getAttribute('style')
-                    if (style && !table.getAttribute('data-user-style')) {
-                        table.setAttribute('data-user-style', style)
-                        table.removeAttribute('style')
-                    }
-                })
-            },
-            toOutput(doc) {
-                doc.querySelectorAll('table').forEach((table) => {
-                    table.removeAttribute('style')
-                    const userStyle = table.getAttribute('data-user-style')
-                    if (userStyle) {
-                        table.setAttribute('style', userStyle)
-                        table.removeAttribute('data-user-style')
-                    }
-                    table.querySelector(':scope > colgroup')?.remove()
-                    table.querySelectorAll('td, th').forEach((cell) => {
-                        const style = cell.getAttribute('style')
-                        if (style) {
-                            const cleaned = style.replace(/min-width\s*:[^;]+;?/gi, '').trim()
-                            if (cleaned) cell.setAttribute('style', cleaned)
-                            else cell.removeAttribute('style')
-                        }
-                        if (cell.getAttribute('colwidth')) cell.removeAttribute('colwidth')
-                        if (cell.getAttribute('colspan') === '1') cell.removeAttribute('colspan')
-                        if (cell.getAttribute('rowspan') === '1') cell.removeAttribute('rowspan')
-                    })
-                })
-            }
-        }
     ]
 }
 
@@ -253,74 +149,6 @@ function getTableContext(tiptap: Editor): { attrs: Record<string, any>; caption:
     }
 
     return { attrs, caption }
-}
-
-function updateCaption(tiptap: Editor, caption: string) {
-    const { $from } = tiptap.state.selection
-
-    let figurePos: number | null = null
-    let figureNode = null
-    let tablePos: number | null = null
-
-    for (let d = $from.depth; d > 0; d--) {
-        const node = $from.node(d)
-        if (node.type.name === 'tableFigure') {
-            figurePos = $from.before(d)
-            figureNode = node
-            break
-        }
-        if (node.type.name === 'table') {
-            tablePos = $from.before(d)
-        }
-    }
-
-    if (figureNode && figurePos !== null) {
-        const firstChild = figureNode.firstChild
-        if (caption) {
-            if (firstChild?.type.name === 'tableCaption') {
-                const from = figurePos + 1
-                const to = from + firstChild.nodeSize
-                tiptap
-                    .chain()
-                    .focus()
-                    .deleteRange({ from, to })
-                    .insertContentAt(from, {
-                        type: 'tableCaption',
-                        content: [{ type: 'text', text: caption }]
-                    })
-                    .run()
-            } else {
-                tiptap
-                    .chain()
-                    .focus()
-                    .insertContentAt(figurePos + 1, {
-                        type: 'tableCaption',
-                        content: [{ type: 'text', text: caption }]
-                    })
-                    .run()
-            }
-        } else if (firstChild?.type.name === 'tableCaption') {
-            const from = figurePos + 1
-            const to = from + firstChild.nodeSize
-            tiptap.chain().focus().deleteRange({ from, to }).run()
-        }
-    } else if (caption && tablePos !== null) {
-        const tableNode = tiptap.state.doc.nodeAt(tablePos)
-        if (!tableNode) return
-        const tableEnd = tablePos + tableNode.nodeSize
-        tiptap
-            .chain()
-            .focus()
-            .deleteRange({ from: tablePos, to: tableEnd })
-            .insertContentAt(tablePos, {
-                type: 'tableFigure',
-                content: [
-                    { type: 'tableCaption', content: [{ type: 'text', text: caption }] },
-                    tableNode.toJSON()
-                ]
-            })
-            .run()
-    }
 }
 
 function openTableDialog(e: TiptapEditor, mode: 'insert' | 'edit') {
